Exportacao JSON Utilizacao de Components Ajuste nas Regras ACL

+++ resources/views/roles/index.blade.php
@@ -1,48 +1,44 @@
 @extends('layouts.app')
 @section('content')
-    <div class="row">
-        <div class="col-lg-12 margin-tb">
-            <div class="pull-left">
-                <h2>Gerenciador de Regras</h2>
-            </div>
-            <div class="pull-right">
-                @can('role-create')
-                    <a class="btn btn-success" href="{{ route('roles.create') }}"> Criar Nova Regra</a>
-                @endcan
-            </div>
-        </div>
-    </div>
-    @if($message = Session::get('success'))
-        <div class="alert alert-success">
-            <p>{{ $message }}</p>
-        </div>
-    @endif
-    <br />
-    <table class="table table-bordered">
-        <tr>
-            <th>Nome</th>
-            <th width="280px">Ação</th>
-        </tr>
-        @foreach($roles as $key => $role)
+    @component('components.titulos-paginas', ["name_class" => "Role", "name_route" => "Roles"])
+        @slot('titulo_pagina')
+            {{ 'Regras' }}
+        @endslot
+    @endcomponent
+    <table id="tabela" class="table table-bordered table-striped tablesorter">
+        <thead>
             <tr>
-                <td>{{ $role->name }}</td>
-                <td>
-                    <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Exibir</a>
-                    @can('role-edit')
-                        <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Editar</a>
-                    @endcan
-                    @can('role-delete')
-                        {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' =>
-                        'display:inline']) !!}
-                        @if($role->name != "Admin")
-                            {!! Form::submit('Excluir', ['class' => 'btn btn-danger']) !!}
-                        @endif
-                        {!! Form::close() !!}
-                    @endcan
-                </td>
+                <th>Nome</th>
+                <th width="280px">Ação</th>
             </tr>
-        @endforeach
+        </thead>
+        <body>
+        @if(count($roles) > 0)
+            @foreach($roles as $key => $role)
+                <tr>
+                    <td>{{ $role->name }}</td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Exibir</a>
+                        @can('role-edit')
+                            <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Editar</a>
+                        @endcan
+                        @can('role-delete')
+                            {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' =>
+                            'display:inline']) !!}
+                            @if($role->name != "Admin")
+                                {!! Form::submit('Excluir', ['class' => 'btn btn-danger']) !!}
+                            @endif
+                            {!! Form::close() !!}
+                        @endcan
+                    </td>
+                </tr>
+            @endforeach
+        @else
+            <tr>
+                <td colspan="2"><em>Não há Regras cadastradas!</em></td>
+            </tr>
+        @endif
+        </body>
     </table>
     {!! $roles->render() !!}
-    <p class="text-center text-primary"><small>Tutorial by rscoder.com</small></p>
 @endsection
