+++ resources/views/usuarios/index.blade.php
@@ -1,22 +1,10 @@
 @extends('layouts.app')
 @section('content')
-    <div class="row">
-        <div class="col-lg-12 margin-tb">
-            <div class="pull-left">
-                <h2>Gestão de Usuários</h2>
-            </div>
-            <div class="pull-right">
-                <a class="btn btn-success" href="{{ route('usuarios.create') }}"> Criar Novo Usuário</a>
-                <a class="btn btn-primary" href="{{ route('exportar', ['id' => 'users']) }}" target="_blank">Exportar Usuários (JSON)</a>
-            </div>
-        </div>
-    </div>
-    @if($message = Session::get('success'))
-        <div class="alert alert-success">
-            <p>{{ $message }}</p>
-        </div>
-    @endif
-    <br />
+    @component('components.titulos-paginas', ["name_class" => "Usuario", "name_route" => "Usuarios"])
+        @slot('titulo_pagina')
+            {{ 'Usuários' }}
+        @endslot
+    @endcomponent
     <table id="tabela" class="table table-bordered table-striped tablesorter">
         <thead>
             <tr>
@@ -27,36 +15,41 @@
             </tr>
         </thead>
         <tbody>
-        @foreach($data as $key => $usuario)
+        @if(count($data) > 0)
+            @foreach($data as $key => $usuario)
+                <tr>
+                    <td>{{ $usuario->name }}</td>
+                    <td>{{ $usuario->email }}</td>
+                    <td>
+                        @php
+                            $tipo_usuario = 'User';
+                        @endphp
+                        @if(!empty($usuario->getRoleNames()))
+                            @foreach($usuario->getRoleNames() as $v)
+                                @php
+                                    $tipo_usuario = $v;
+                                @endphp
+                                <label class="badge badge-success">{{ $v }}</label>
+                            @endforeach
+                        @endif
+                    </td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('usuarios.show', $usuario->id) }}">Exibir</a>
+                        <a class="btn btn-primary" href="{{ route('usuarios.edit', $usuario->id) }}">Editar</a>
+                        {!! Form::open(['method' => 'DELETE', 'route' => ['usuarios.destroy', $usuario->id], 'style' => 'display:inline']) !!}
+                        @if(Auth::user()->id!=$usuario->id && $tipo_usuario != "Admin")
+                            {!! Form::submit('Excluir', ['class' => 'btn btn-danger']) !!}
+                        @endif
+                        {!! Form::close() !!}
+                    </td>
+                </tr>
+            @endforeach
+        @else
             <tr>
-                <td>{{ $usuario->name }}</td>
-                <td>{{ $usuario->email }}</td>
-                <td>
-                    @php
-                        $tipo_usuario = 'User';
-                    @endphp
-                    @if(!empty($usuario->getRoleNames()))
-                        @foreach($usuario->getRoleNames() as $v)
-                            @php
-                                $tipo_usuario = $v;
-                            @endphp
-                            <label class="badge badge-success">{{ $v }}</label>
-                        @endforeach
-                    @endif
-                </td>
-                <td>
-                    <a class="btn btn-info" href="{{ route('usuarios.show', $usuario->id) }}">Exibir</a>
-                    <a class="btn btn-primary" href="{{ route('usuarios.edit', $usuario->id) }}">Editar</a>
-                    {!! Form::open(['method' => 'DELETE', 'route' => ['usuarios.destroy', $usuario->id], 'style' => 'display:inline']) !!}
-                    @if(Auth::user()->id!=$usuario->id && $tipo_usuario != "Admin")
-                        {!! Form::submit('Excluir', ['class' => 'btn btn-danger']) !!}
-                    @endif
-                    {!! Form::close() !!}
-                </td>
+                <td colspan="4"><em>Não há Usuários cadastrados!</em></td>
             </tr>
-        @endforeach
+        @endif
         </tbody>
     </table>
     {!! $data->render() !!}
-    <p class="text-center text-primary"><small>Tutorial by rscoder.com</small></p>
 @endsection
